// IonicThesis/src/app/models/current-balance.ts
export interface CurrentBalance {
    huf?: number, //forint
    eur?: number, //euro
    usd?: number, //amerikai dollár
    jpy?: number, //japán yen
    gbp?: number, //angol font
    chf?: number, //svájci frank
    aud?: number, //ausztrál dollár
    cad?: number, //kanadai dollár
    cny?: number, //kínai juan
    bgn?: number, //bolgár leva
    czk?: number, //cseh korona
    dkk?: number, //dán korona
    hrk?: number, //horvát kuna
    mxn?: number, //mexikói peso
    nok?: number, //norvég korona
    nzd?: number, //új-zélandi dollár
    pln?: number, //lengyel zloty
    ron?: number, //román lej
    rsd?: number, //szerb dinár
    rub?: number, //orosz rubel
    sek?: number, //svéd korona
    try?: number, //török líra
    uah?: number, //ukrán hrivnya
    zar?: number, //dél-afrikai rand
    balanceHUF?: number,
    expenseAmountHUF?: number,
    incomeAmountHUF?: number,
}